makefile + operation

+++ libft/ft_atoi.c
@@ -13,11 +13,11 @@
 #include <stdlib.h>
 #include "libft.h"
 
-int	ft_atoi(const char *nptr)
+long	ft_atoi(const char *nptr)
 {
 	int	i;
 	int	neg;
-	int	res;
+	long	res;
 
 	i = 0;
 	neg = 1;
